+++ depsi/constants.py
@@ -6,3 +6,6 @@ EARTH_RADIUS = 6378136
 
 # Speed of light in vacuum
 SPEED_OF_LIGHT = 299792458.0  # m/s
+
+# Sentinel-1 wavelength
+WAVELENGTH_S1 = 0.055465763  # m
