+++ depsi/stochastic.py
