--- network.py
@@ -1,2054 +1,0 @@
-"""Module for creating networks from STM points."""
-
-import logging
-import math
-from typing import Literal
-
-import matplotlib.pyplot as plt
-import networkx as nx
-import numpy as np
-import scipy
-import sparse
-import xarray as xr
-from scipy.spatial import Delaunay, KDTree, distance_matrix
-from sklearn.neighbors import BallTree
-
-from depsi.arc_estimation import arc_estimation_control_network
-from depsi.constants import EARTH_RADIUS
-from depsi.stats import pretest
-from depsi.utils import compute_phase_difference, get_distance
-
-logger = logging.getLogger(__name__)
-
-
-# Constants for MHT in network integration
-ALPHA0 = 0.1  # Significance level for 1-dimensional test
-GAMMA0 = 0.5  # Power of the test
-# Overall Model Test threshold for accepting the network
-# In arc/point rejection phase, if OMT < OMT_THRES, stop rejection iteration
-# In ambiguity fixing phase, if OMT < OMT_THRES, stop fixing iteration
-# In arc/point rejection phase this is hardly triggered
-OMT_THRES = 1e-10
-# Threshold for arc rejection statistics TT1
-# If for all arcs max(TT1) < TT1_THRES, stop rejection iteration
-# For most cases this threshold is triggered in rejection phase
-TT1_THRES = 1.0
-
-
-def spatial_integration(
-    stm_pnts: xr.Dataset,
-    stm_arcs: xr.Dataset,
-    key_sdphase: str = "sd_phase",
-    key_arc_quality: str = "temp_coh",
-    threshold_arc_quality: float = 0.5,
-    idx_refpnt: int | None = None,
-    min_arc_connections: int = 3,
-    largest_component_ratio: float = 0.8,
-    parallel: bool = False,
-    sparse_mode: bool = False,
-    ensure_network_while_mht: bool = False,
-    arc_estimation_method: Literal["periodogram"] = "periodogram",
-    skip_network_adaptation: bool = False,
-    max_iterations_adaptation: int = None,
-) -> tuple[xr.Dataset, xr.Dataset]:
-    """Spatially integrate the ambiguities of network arcs to points.
-
-    This function estimates the integer ambiguities of the points from arc ambiguities. It assumes a network
-    has been formed from the points by calling "network.form_network", and arc ambiguities have been estimated
-    by calling relevant functions in the "depsi.arc_estimation" module.
-
-    The function returns an updated STM of arcs which contains the adjusted arc ambiguities after spatial integration,
-    as well as an updated STM of points which contains the estimated point ambiguities.
-
-    The following steps are performed:
-    1. Validate the network arcs (`stm_arcs`) against the point STM (`stm_pnts`)
-    2. Select arcs based on quality threshold and ensure minimum connections for all points
-    3. Select a reference point which assumes zero phase (hence zero ambiguity)
-    4. Adjust the network by removing arcs/points which potentially cause errors using Multi-Hypothesis Testing (MHT)
-    5. Adjust the ambiguities per time epoch to make sure spatial solutions give zero residuals.
-    6. Calculate point ambiguities and unwrapped phases w.r.t. the reference point.
-
-    Parameters
-    ----------
-    stm_pnts : xr.Dataset
-        Space-Time Matrix of points.
-    stm_arcs : xr.Dataset
-        Space-Time Matrix of arcs.
-        The arcs should be formed from stm_pnts, using the "network.form_network" function. This function generates
-        coordinates "source" and "target" in stm_arcs which refer to the indices in stm_pnts. An "uid" data variable
-        is also generated by "form_network" for easy indexing.
-        An arc estimation should have been applied on stm_arcs before calling "spatial_integration". One can use
-        relevant functions in "depsi.arc_estimation" module for this purpose. Arc estimation adds the variable
-        "ambiguities" to stm_arcs, which are the estimated arc ambiguities. It also adds quality variables such as
-        "temp_coh" (ensemble coherence), which are used to select arcs for spatial integration.
-    key_sdphase : str, optional
-        Key of the single difference phase variable in stm_pnts, by default "sd_phase"
-        This phase is used to compute unwrapped phases after ambiguity estimation.
-    key_arc_quality : str, optional
-        Key of the arc quality variable in stm_arcs, by default "temp_coh"
-    threshold_arc_quality : float, optional
-        Threshold for arc quality, by default 0.5
-    idx_refpnt : int | None, optional
-        Index of the reference point in stm_pnts. If None, the source point of the arc with highest quality is selected
-        as the reference point.
-    min_arc_connections : int, optional
-        Minimum number of connections for arcs, by default 3
-    largest_component_ratio : float, optional
-        Threshold for determining the largest component when multiple components exist in the network, by default 0.8.
-        When removing arcs/points, it may happen that the network is split into multiple disconnected components.
-        In this case, only the largest component is kept and the others are discarded.
-        The largest component should contain at least this fraction of the total points, otherwise an error is raised.
-        Hence, when an error should be raised at all times, set this value to 1.0 .
-    parallel : bool, optional
-        Whether to use parallel processing, by default False
-    sparse_mode : bool, optional
-        Whether to use sparse matrix format for large networks, by default False
-    ensure_network_while_mht : bool, optional
-        Whether to ensure minimum connections in MHT network adaptation, by default False
-    arc_estimation_method : Literal["periodogram"], optional
-        Method used for arc estimation, by default "periodogram".
-        This constrains the method used for VCM computation.
-    skip_network_adaptation : bool, optional
-        Whether to skip network adaptation by MHT, by default False.
-        When enabling this option, it is recommended to set the threshold_arc_quality to a
-        high value (e.g. 0.75) to ensure only high-quality arcs are selected for spatial
-        integration.
-    max_iterations_adaptation : int, optional
-        Maximum number of iterations for network adaptation.
-        If None, the maximum number if iterations will be the number of arcs.
-
-    Returns
-    -------
-    xr.Dataset, xr.Dataset
-        Updated Space-Time Matrix of arcs and updated Space-Time Matrix of points.
-        For arcs, the "ambiguities" variable contains the adjusted arc ambiguities after spatial integration.
-        For points, the "ambiguities" variable contains the estimated point ambiguities, and "unwrapped_phase"
-        contains the unwrapped phase w.r.t. the reference point.
-
-    References
-    ----------
-    Van Leijen, F.J.. "Persistent scatterer interferometry based on geodetic estimation theory." (2014).
-    """
-    # Check parallelization behavior
-    if parallel:
-        raise NotImplementedError("Dask support is not implemented yet for spatial_integration.")
-    else:
-        # Compute all data into memory
-        stm_pnts = stm_pnts.compute()
-        stm_arcs = stm_arcs.compute()
-
-    # Validate that stm_arcs are formed from stm_pnts
-    if (stm_arcs["source"].max().values >= stm_pnts.sizes["space"]) or (
-        stm_arcs["target"].max().values >= stm_pnts.sizes["space"]
-    ):
-        raise ValueError("stm_arcs contains source/target indices that exceed the number of points in stm_pnts.")
-    if "ambiguities" not in stm_arcs:
-        raise ValueError("stm_arcs does not contain 'ambiguities' variable. Please estimate arc ambiguities first.")
-    if key_arc_quality not in stm_arcs:
-        raise ValueError(f"stm_arcs does not contain '{key_arc_quality}' variable for arc quality assessment.")
-
-    # Check arc estimation method, this constrains VCM computation method
-    if arc_estimation_method not in ["periodogram"]:
-        raise NotImplementedError(f"Unknown arc estimation method {arc_estimation_method}.")
-
-    if sparse_mode:
-        raise NotImplementedError("Sparse mode is not implemented yet for spatial_integration.")
-
-    # If idx_refpnt is specified
-    # Get radar coordinates of the reference point before any shape change
-    if idx_refpnt is not None:
-        azimuth_refpnt = stm_pnts["azimuth"].isel(space=idx_refpnt).values
-        range_refpnt = stm_pnts["range"].isel(space=idx_refpnt).values
-
-    # Select arcs with quality > threshold_arc_quality
-    # Then ensure all points have at least min_arc_connections connections
-    mask = (np.abs(stm_arcs[key_arc_quality]) > threshold_arc_quality).compute()
-    stm_arcs = stm_arcs.where(mask, drop=True)
-    stm_arcs, stm_pnts = _ensure_network_min_connections(stm_arcs, stm_pnts, min_arc_connections)
-
-    # Ensure the network is a single connected component after arc selection and point removal
-    stm_arcs, stm_pnts = _ensure_single_network(stm_arcs, stm_pnts, largest_component_ratio)
-
-    # Select reference point as the source pnt of arcs with highest temp_coh
-    if idx_refpnt is None:
-        idx_arc_max_coh = stm_arcs[key_arc_quality].argmax().values
-        idx_refpnt = stm_arcs["source"].isel(space=idx_arc_max_coh).values
-        azimuth_refpnt = stm_pnts["azimuth"].isel(space=idx_refpnt).values
-        range_refpnt = stm_pnts["range"].isel(space=idx_refpnt).values
-    else:
-        # Make sure idx_refpnt is still valid after arc selection and point removal
-        mask_refpnt = (stm_pnts["azimuth"].values == azimuth_refpnt) & (stm_pnts["range"].values == range_refpnt)
-        if not np.any(mask_refpnt):
-            raise ValueError(
-                f"Reference point ({azimuth_refpnt}, {range_refpnt}) removed after arc selection. "
-                f"Please choose another reference point."
-            )
-
-        # Get reference point from radar coordinates
-        idx_refpnt = np.where(mask_refpnt)[0][0]
-
-    # Adjust the network by removing bad arcs/points using MHT
-    if skip_network_adaptation:
-        logger.info("Skipping MHT network adaptation step.")
-        stm_arcs_adjusted, stm_pnts_adjusted = stm_arcs, stm_pnts
-    else:
-        stm_arcs_adjusted, stm_pnts_adjusted = _mht_network_adaptation(
-            stm_arcs,
-            stm_pnts,
-            idx_refpnt,
-            azimuth_refpnt,
-            range_refpnt,
-            ensure_network_while_mht,
-            sparse_mode,
-            arc_estimation_method,
-            max_iterations_adaptation,
-            largest_component_ratio,
-        )
-
-        # Update idx_refpnt after MHT adaptation
-        idx_refpnt = np.where(
-            (stm_pnts_adjusted["azimuth"].values == azimuth_refpnt)
-            & (stm_pnts_adjusted["range"].values == range_refpnt)
-        )[0][0]
-
-    # Adjust ambiguities to fix unwrapping errors
-    stm_arcs_output, stm_pnts_output, idx_refpnt = _ambiguity_adaptation(
-        stm_arcs_adjusted,
-        stm_pnts_adjusted,
-        idx_refpnt,
-        sparse_mode,
-        arc_estimation_method,
-    )
-
-    # Assign idx_refpnt as attribute to stm_pnts_output
-    stm_pnts_output = stm_pnts_output.assign_attrs({"idx_refpnt": idx_refpnt})
-
-    # Add unwrapped phase to stm_pnts_output
-    # Unwrapped phase is w.r.t. the reference point
-    # Therefore the sd_phase of the reference point is subtracted
-    unwrapped_phase_pnts = (
-        stm_pnts_output[key_sdphase].data
-        + stm_pnts_output["ambiguities"].data * 2 * np.pi
-        - np.tile(
-            stm_pnts_output[key_sdphase].isel(space=idx_refpnt).data,
-            (stm_pnts_output.sizes["space"], 1),
-        )
-    )
-    stm_pnts_output["unwrapped_phase"] = (("space", "time"), unwrapped_phase_pnts)
-
-    return stm_arcs_output, stm_pnts_output
-
-
-def form_network(
-    stm: xr.Dataset,
-    key_phase: str,
-    key_h2ph: str,
-    key_Btemporal: str,
-    key_complex: str = "complex",
-    key_xcrds: str = "lon",
-    key_ycrds: str = "lat",
-    network_method: Literal["redundant", "delaunay"] = "redundant",
-    max_length: float = None,
-    min_links: int = 16,
-    num_partitions: int = 8,
-    dphase_method: Literal["conjmult", "subtract"] = "subtract",
-) -> xr.Dataset:
-    """Generate an STM of arcs from an STM of points.
-
-    Parameters
-    ----------
-    stm : xr.Dataset
-        Space-Time Matrix of scatterers.
-    key_phase : str
-        Key of the phase values in the STM.
-        This phase will be used to compute the differential arc phase.
-    key_h2ph : str
-        Key of the h2ph values in the STM.
-        The arc h2ph will be computed as the average between source and target.
-    key_Btemporal : str
-        Key of the temporal baseline values in the STM.
-    key_complex : str, optional
-        Key of the complex values, by default "complex"
-    key_xcrds  : str, optional
-        Key of the x coordinates for calulating arc length, by default "lon"
-    key_ycrds  : str, optional
-        Key of the y coordinates for calulating arc length, by default "lat"
-    network_method : Literal["redundant", "delaunay"], optional
-        network formation method, by default "redundant"
-    max_length : float, optional
-        maximum arc length, by default None
-    min_links : int, optional
-        minimum links per point, by default 16
-        only effective when network_method is "redundant"
-    num_partitions : int, optional
-        number of partitions of searching when forming redundant network, by default 8
-        only effective when network_method is "redundant"
-    dphase_method : Literal["conjmult", "subtract"], optional
-        method of computing phase difference, by default "subtract"
-        "subtract" method subtracts the source phase from the target phase (without re-wrapping);
-        "conjmult" method computes the phase difference by conjugate multiplication:
-            d_phase = np.angle(complex_target * complex_source.conj())
-
-    Returns
-    -------
-    xr.Dataset
-        Space-Time Matrix of arcs, containing the following variables:
-        - d_phase: the arc phase, which is the difference between source and target points
-        - Btemp: the temporal baseline, which is the same for all arcs
-        - h2ph: the arc h2ph, which is the average between source and target points
-    """
-    # Generate the network arcs.
-    if network_method == "redundant":
-        if min_links <= 0:
-            logger.error(f"min_links must be strictly positive (currently: {min_links})")
-            return
-        if num_partitions <= 0:
-            logger.error(f"num_partitions must be strictly positive (currently: {num_partitions})")
-            return
-    elif network_method != "delaunay":
-        raise NotImplementedError(f"Unknown network method {network_method}, known are delaunay and redundant")
-
-    # Collect point coordinates.
-    indices = [stm[coord] for coord in [key_xcrds, key_ycrds]]
-    coordinates = np.column_stack(indices)
-
-    arcs = None
-
-    # Create network arcs as list of tuples of point ids.
-    if network_method == "delaunay":
-        arcs = _generate_arcs_delaunay(coordinates, max_length)
-    elif network_method == "redundant":
-        arcs = _generate_arcs_redundant(coordinates, max_length, min_links, num_partitions)
-
-    # Compute the phase difference.
-    arcs_unzipped = list(zip(*arcs, strict=False))
-    source_idx = list(arcs_unzipped[0])
-    target_idx = list(arcs_unzipped[1])
-
-    if dphase_method not in ["conjmult", "subtract"]:
-        raise NotImplementedError(f"Unknown dphase_method '{dphase_method}'.")
-    dict_key_method = {"subtract": key_phase, "conjmult": key_complex}  # mapping for selecting the correct key
-    d_phase = compute_phase_difference(
-        stm.isel(space=source_idx)[dict_key_method[dphase_method]].data,
-        stm.isel(space=target_idx)[dict_key_method[dphase_method]].data,
-        method=dphase_method,
-    )
-
-    # Temporal baseline
-    Btemp = stm[key_Btemporal].data
-
-    # Height to phase factor
-    h2ph = (stm[key_h2ph].isel(space=source_idx).data + stm[key_h2ph].isel(space=target_idx).data) / 2
-
-    # Generate a unique identifier of arcs based on source and target for easy indexing
-    # This is because when updating network, points can be removed and reindexed
-    # Therefore we cannot use 2d index (source, target) as uid
-    # NOTE: This encoding is safe for typical networks (<10,000 points, <100,000 arcs).
-    # For 10,000 points, max UID ≈ 1e9, well within int64 max (2^63-1 ≈ 9.22e18), providing
-    # a safety margin of >9 billion times. Even for 1M points, safety margin is >900,000x.
-    scale = 10 ** (math.floor(math.log10(stm.sizes["space"])) + 1)  # Scale to ensure no overlap
-    uid = scale * (np.array(source_idx) + 1) + (np.array(target_idx) + 1)  # Plus one to avoid zero uid
-    uid = uid.astype(np.int64)
-
-    arcs = xr.Dataset(
-        data_vars={
-            "d_phase": (["space", "time"], d_phase),
-            "Btemp": (["time"], Btemp),
-            "h2ph": (["space", "time"], h2ph),
-        },
-        coords={"source": (["space"], source_idx), "target": (["space"], target_idx), "uid": (["space"], uid)},
-        attrs=stm.attrs,
-    )
-
-    return arcs
-
-
-def _mht_network_adaptation(
-    stm_arcs: xr.Dataset,
-    stm_pnts: xr.Dataset,
-    idx_refpnt: int,
-    azimuth_refpnt: int | float,
-    range_refpnt: int | float,
-    ensure_network_while_mht: bool,
-    sparse_mode: bool,
-    arc_estimation_method: str,
-    max_iterations_adaptation: int,
-    largest_component_ratio: float,
-) -> tuple[xr.Dataset, xr.Dataset]:
-    """Adjust the network by removing bad arcs/points by applying MHT.
-
-    This function implements the Multi-Hypothesis Testing (MHT) approach iteratively to identify and remove
-    arcs and points to reduce the overall residual in ambiguity estimation.
-
-    Parameters
-    ----------
-    stm_arcs : xr.Dataset
-        Space-Time Matrix of arcs.
-    stm_pnts : xr.Dataset
-        Space-Time Matrix of points.
-    idx_refpnt : int
-        Index of the reference point in stm_pnts.
-    azimuth_refpnt : int | float
-        Azimuth coordinate of the reference point.
-    range_refpnt : int | float
-        Range coordinate of the reference point.
-    ensure_network_while_mht : bool
-        Whether to ensure minimum connections in MHT network adaptation.
-    sparse_mode : bool
-        Whether to use sparse matrix format for large networks.
-    arc_estimation_method : str
-        Method used for arc estimation.
-    max_iterations_adaptation : int
-        Maximum number of iterations for network adaptation.
-    largest_component_ratio : float
-        Threshold for determining the largest component when multiple components exist in the network.
-
-    Returns
-    -------
-    xr.Dataset, xr.Dataset
-        Updated Space-Time Matrix of arcs and points.
-    """
-    # Setup functional and stochastic model
-    A = _network_relation_matrix(
-        stm_arcs["source"], stm_arcs["target"], stm_pnts.sizes["space"], idx_refpnt, sparse_mode
-    )  # Network relation matrix A
-
-    if arc_estimation_method == "periodogram":
-        Qyy_diag = 1 - stm_arcs["temp_coh"].values
-    else:
-        raise NotImplementedError(f"arc_estimation_method '{arc_estimation_method}' is not supported.")
-    invQy = np.diag(1 / Qyy_diag)
-
-    _, echeck, _ = _solve_float_ambiguities(A, stm_arcs["ambiguities"].data, invQy)  # Estimate initial residual
-    OMT = np.diag(echeck.T @ invQy @ echeck).sum()  # Test statistics for Overall Model Test
-
-    # Setup test parameters
-    kb_dict = {}
-    max_con = np.abs(A).sum(axis=0).max()
-    for n_con in range(1, max_con + 1):
-        _, k1, kb, _ = pretest(n_con, ALPHA0, GAMMA0)
-        kb_dict[n_con] = kb
-
-    # By default, set max_iterations_adaptation to the number of arcs
-    if max_iterations_adaptation is None:
-        max_iterations_adaptation = stm_arcs.sizes["space"]
-
-    # Iteratively remove arcs/points until OMT and all arc statistics pass the test
-    stm_pnts_updated = stm_pnts.copy()
-    stm_arcs_updated = stm_arcs.copy()
-    TT1max = TT1_THRES + 1.0  # Initial TT1_max to trigger the while loop
-    niter = 0
-    while (OMT >= OMT_THRES) and (TT1max >= TT1_THRES) and (niter < max_iterations_adaptation):
-        # The iteration stops when one of the following conditions is met:
-        # 1) overall model test pass: OMT < OMT_THRES (very rare case)
-        # 2) all arc test statistics smaller than threshold: max(TT1) < TT1_THRES (most common case)
-        # 3) maximum number of iterations reached (fail case)
-        logger.debug(f"MHT iteration {niter}: OMT={OMT:.2e}")
-
-        # Because OMT failed, choose from two Ha: 1) remove an arc; 2) remove a point
-        # Decision is made based on flag_rm
-        flag_rm, idx_rm, TT1max, TTqmax = _mht_network_adaptation_reject_one(
-            A, stm_arcs_updated["ambiguities"].data, Qyy_diag, k1, kb_dict
-        )
-
-        if flag_rm == 0:  # remove arc
-            logger.debug(f"MHT iteration {niter}: removing arc index {idx_rm} with TT1={TT1max:.2f}")
-            stm_arcs_updated = stm_arcs_updated.drop_isel(space=idx_rm)  # Remove the arc
-        elif flag_rm == 1:  # remove point
-            if idx_rm >= idx_refpnt:
-                idx_rm += 1  # Adjust index due to removed reference point column in A
-
-            logger.debug(f"MHT iteration {niter}: removing point index {idx_rm} with TT1={TT1max:.2f}")
-
-            # Removing points is achieved by removing all arcs connects to the point
-            # Later the points will be actually removed when ensuring minimum connections
-            # Arc indices connecting to the point to remove
-            idx_arcs_selected = np.where(
-                ((stm_arcs_updated["source"] != idx_rm) & (stm_arcs_updated["target"] != idx_rm)).data
-            )[0]
-            # Remove all arcs connects to the point to remove
-            stm_arcs_updated = stm_arcs_updated.isel(space=idx_arcs_selected)
-
-        if ensure_network_while_mht:
-            min_connections_to_ensure = 3  # Ensure all points in the network have at least 3 connections
-        else:
-            min_connections_to_ensure = 2  # Just ensure all points are connected in the network
-
-        # This makes sure all points can be tested in case of disagreement between arcs
-        stm_arcs_updated, stm_pnts_updated = _ensure_network_min_connections(
-            stm_arcs_updated, stm_pnts_updated, min_connections=min_connections_to_ensure
-        )
-
-        # Ensure the network is a single connected component after arc/point removal
-        stm_arcs_updated, stm_pnts_updated = _ensure_single_network(
-            stm_arcs_updated, stm_pnts_updated, largest_component_ratio
-        )
-
-        # Make sure the reference point is still in stm_pnts_updated, by checking its azimuth and range
-        mask_refpnt = (stm_pnts_updated["azimuth"].values == azimuth_refpnt) & (
-            stm_pnts_updated["range"].values == range_refpnt
-        )
-        if not np.any(mask_refpnt):
-            raise ValueError(
-                f"Reference point ({azimuth_refpnt}, {range_refpnt}) removed in the MHT process. "
-                f"Please choose another reference point."
-            )
-        idx_refpnt = np.where(mask_refpnt)[0][0]  # Update idx_refpnt
-
-        # Get indices of selected arcs based on uid
-        if arc_estimation_method == "periodogram":
-            Qyy_diag = 1 - stm_arcs_updated["temp_coh"].values  # VCM diagonal from temporal coherence
-        invQy = np.diag(1 / Qyy_diag)
-
-        A = _network_relation_matrix(
-            stm_arcs_updated["source"],
-            stm_arcs_updated["target"],
-            stm_pnts_updated.sizes["space"],
-            idx_refpnt,
-            sparse_mode,
-        )  # Update A matrix
-        _, echeck, _ = _solve_float_ambiguities(
-            A, stm_arcs_updated["ambiguities"].data, invQy
-        )  # Estimate residual again
-        OMT = np.diag(echeck.T @ invQy @ echeck).sum()  # Update OMT statistic
-
-        niter += 1
-
-    if niter >= stm_arcs.sizes["space"]:
-        raise RuntimeError(
-            "Maximum number of iterations reached in MHT network adaptation. "
-            "The network may still contain bad arcs or points."
-        )
-
-    return stm_arcs_updated, stm_pnts_updated
-
-
-def _mht_network_adaptation_reject_one(
-    A: np.ndarray | scipy.sparse.spmatrix,
-    y: np.ndarray,
-    Qyy_diag: np.ndarray,
-    k1: float,
-    kb_dict: dict,
-) -> tuple[int, int, float, float]:
-    """Remove one point/arc from the network to reduce the residual in ambiguity estimation."""
-    # Retrieve shapes
-    N_arcs, N_epochs = y.shape
-    N_points = A.shape[1]
-
-    # Inverse of VCM of observations
-    if Qyy_diag.ndim == 1:  # Diagonal VCM
-        invQy = np.diag(1 / Qyy_diag)
-        Qyy = np.diag(Qyy_diag)
-    else:
-        raise NotImplementedError("Currently only diagonal VCM is supported. Qyy_diag should be a 1d array.")
-
-    # Solve ambiguities as float
-    _, echeck, invAtQyA = _solve_float_ambiguities(A, y, invQy)
-
-    # Post-priori VCM of residuals
-    Qecheck = Qyy - (A @ invAtQyA @ A.T)  # TODO: check how to handle large Qecheck
-
-    # Test statistics TT1 per arc
-    Qecheck_diag = np.array(np.diag(Qecheck).flatten()).squeeze()
-    w = echeck**2 / np.tile(np.abs(Qecheck_diag), (N_epochs, 1)).T
-    TT1 = np.sum(w, axis=1) / k1**2
-    TT1max = max(TT1)
-
-    # Test statistics per point
-    # Build arc-point connectivity mask and
-    connected_mask = A != 0
-    # Drop one connected arc per point to create the basis (see e.g. verhoef97).
-    has_connection = connected_mask.any(axis=0)
-    first_connected_idx = np.argmax(connected_mask, axis=0)
-    selected_mask = connected_mask.copy()
-    selected_mask[first_connected_idx[has_connection], np.where(has_connection)[0]] = False
-    # Tq for point q: sum_i(sum_t(e_i,t^2) / Qe_i) over selected arcs i connected to point q.
-    e2_sum = np.sum(echeck**2, axis=1)
-    e2_sum_weighted = e2_sum / np.abs(Qecheck_diag)
-    Tq_num = selected_mask.T @ e2_sum_weighted
-    # Find where to calculate TTq based on connectivity
-    narcs_connected = selected_mask.sum(axis=0).astype(int)
-    kb_vals = np.array([kb_dict.get(d, np.nan) for d in narcs_connected], dtype=float)
-    valid = (narcs_connected > 0) & np.isfinite(kb_vals) & (kb_vals != 0)
-    # Calculate TTq for points with valid kb values
-    # assign -inf to invalid ones to make sure they won't be selected for removal
-    TTq = np.full(N_points, -np.inf, dtype=float)
-    with np.errstate(divide="ignore", invalid="ignore"):
-        TTq[valid] = Tq_num[valid] / kb_vals[valid]
-    TTqmax = max(TTq)
-
-    # Decision one removal strategy
-    if TT1max > TTqmax:
-        idx_removal = np.argmax(TT1)  # index of arc to remove
-        flag_removal = 0  # remove arc
-    else:
-        idx_removal = np.argmax(TTq)  # index of point to remove
-        flag_removal = 1  # remove point
-
-    return flag_removal, idx_removal, TT1max, TTqmax
-
-
-def _ambiguity_adaptation(
-    stm_arcs: xr.Dataset,
-    stm_pnts: xr.Dataset,
-    idx_refpnt: int,
-    sparse_mode: bool,
-    arc_estimation_method: str,
-) -> tuple[xr.Dataset, xr.Dataset, int]:
-    """Fix unwrapping errors by adjusting ambiguities per epoch.
-
-    This function iterates over each epoch and adjusts the ambiguities to make sure the spatial
-    solutions of ambiguities gives zero residuals for each epoch.
-
-    Parameters
-    ----------
-    stm_arcs : xr.Dataset
-        Space-Time Matrix of arcs.
-    stm_pnts : xr.Dataset
-        Space-Time Matrix of points.
-    idx_refpnt : int
-        Index of the reference point.
-    sparse_mode : bool
-        Whether to use sparse matrix format for large networks.
-    arc_estimation_method : str
-        Method used for arc estimation.
-
-    Returns
-    -------
-    (xr.Dataset, xr.Dataset)
-        Updated Space-Time Matrices of arcs and points.
-        For arcs, the "ambiguities" variable contains the adjusted arc ambiguities.
-        For points, the "ambiguities" variable are estimated from the adjusted arc ambiguities.
-    """
-    # Setup functional and stochastic model
-    A = _network_relation_matrix(
-        stm_arcs["source"], stm_arcs["target"], stm_pnts.sizes["space"], idx_refpnt, sparse_mode
-    )
-    if arc_estimation_method == "periodogram":
-        Qyy_diag = 1 - stm_arcs["temp_coh"].values  # VCM diagonal from temporal coherence
-    else:
-        raise NotImplementedError(f"arc_estimation_method '{arc_estimation_method}' is not supported in form_network.")
-    invQy = np.diag(1 / Qyy_diag)
-
-    # Initialize adjusted ambiguities storage, shape: (n_points-1, n_epochs)
-    # Space dimension is n_points-1 because reference point is excluded
-    acheck = np.zeros((stm_pnts.sizes["space"] - 1, stm_pnts.sizes["time"]))
-
-    # Fix unwrapping ambiguities by looping over epochs
-    stm_arcs_updated = stm_arcs.copy()
-    stm_pnts_updated = stm_pnts.copy()
-    for epoch in range(stm_pnts.sizes["time"]):
-        logger.debug(f"Adjusting ambiguities for epoch {epoch}")
-        y = stm_arcs["ambiguities"].isel(time=epoch).data
-        acheck_ifg, echeck_ifg, _ = _solve_float_ambiguities(A, y, invQy)
-        OMT = echeck_ifg.T @ invQy @ echeck_ifg
-        idx_previous_arc_fix = -1  # Avoid fixing the same arc again in the same epoch
-
-        while OMT >= OMT_THRES:  # While the OMT fails, fix for this epoch
-            # Find arc index with largest abs echeck
-            # When OMT > kOMT, echeck_ifg[idx_max_echeck] is guaranteed to be non-zero
-            idx_sort = np.argsort(np.abs(echeck_ifg))[::-1]  # Indices of echeck sorted by abs value, descending
-            idx_max_echeck = idx_sort[0]  # Index of arc with largest abs echeck
-            if idx_max_echeck == idx_previous_arc_fix:
-                # If get same arc as previous fix, take the second largest
-                idx_max_echeck = idx_sort[1]
-
-            if np.round(abs(echeck_ifg[idx_max_echeck])) >= 1:  # If >= 1, minus closest integer
-                y[idx_max_echeck] -= np.round(echeck_ifg[idx_max_echeck])
-            elif echeck_ifg[idx_max_echeck] > 0:  # if (0, 1), minus 1
-                y[idx_max_echeck] -= 1.0
-            elif echeck_ifg[idx_max_echeck] < 0:  # if (-1, 0), plus 1
-                y[idx_max_echeck] += 1.0
-
-            idx_previous_arc_fix = idx_max_echeck  # record the fixed arc index
-
-            # Recalculate OMT
-            acheck_ifg, echeck_ifg, _ = _solve_float_ambiguities(A, y, invQy)
-            OMT = echeck_ifg.T @ invQy @ echeck_ifg
-
-            logger.debug(f"Fixing arc index {idx_max_echeck}, new OMT={OMT:.2e}")
-
-        stm_arcs_updated["ambiguities"][:, epoch] = y  # Store adjusted arc ambiguities
-        acheck[:, epoch] = acheck_ifg  # Store adjusted point ambiguities
-
-    # Round acheck to closest integer
-    acheck = np.round(acheck).astype(np.int16)
-
-    # Assign reference point ambiguities as zero
-    acheck_full = np.zeros((stm_pnts.sizes["space"], stm_pnts.sizes["time"])).astype(np.int16)
-    acheck_full[np.arange(stm_pnts.sizes["space"]) != idx_refpnt, :] = acheck
-
-    # Assign acheck_full to stm_pnts_updated
-    stm_pnts_updated["ambiguities"] = (["space", "time"], acheck_full)
-
-    return stm_arcs_updated, stm_pnts_updated, idx_refpnt
-
-
-def _ensure_network_min_connections(
-    stm_arcs: xr.Dataset,
-    stm_pnts: xr.Dataset,
-    min_connections: int,
-) -> tuple[xr.Dataset, xr.Dataset]:
-    """Ensure that all points in the network have at least min_connections arcs.
-
-    This is achieved by an iterative process of removing points which have less than
-    min_connections arcs, and removing all arcs connected to these points.
-    The process is repeated until no more points are removed.
-
-    Parameters
-    ----------
-    stm_arcs : xr.Dataset
-        Space-Time Matrix of arcs.
-    stm_pnts : xr.Dataset
-        Space-Time Matrix of points.
-    min_connections : int
-        Minimum number of connections for each point.
-
-    Returns
-    -------
-    xr.Dataset, xr.Dataset
-        Updated Space-Time Matrix of arcs and points.
-    """
-    # Ensure all points have at least min_connections connections
-    previous_size = -1  # Initialize with an impossible value to trigger the while loop
-    # Keep iterating until no more points are removed
-    while stm_pnts.sizes["space"] != previous_size:
-        previous_size = stm_pnts.sizes["space"]
-        # Remove points with < min_connections connections
-        stm_pnts, stm_arcs = _remove_network_points_min_connections(stm_pnts, stm_arcs, min_connections)
-
-    return stm_arcs, stm_pnts
-
-
-def _ensure_single_network(
-    stm_arcs: xr.Dataset, stm_pnts: xr.Dataset, largest_component_ratio: float
-) -> tuple[xr.Dataset, xr.Dataset]:
-    """Ensure the network is connected and discard the smaller disconnected sub-network(s).
-
-    This function utilizes the NetworkX library to identify connected components in the network
-    formed by stm_arcs and stm_pnts. When building the graph, the point indices are used as node identifiers,
-    and the "source" and "target" coordinates in stm_arcs are used to add edges between the corresponding nodes.
-    Note that "source" and "target" coordinates in stm_arcs are indices of the points STM stm_pnts,
-    but not necessarily the same as the "space" coordinate of stm_pnts.
-
-    If there are multiple connected components, only the largest one is kept and the others are discarded.
-    However, if the largest component is smaller than a certain ratio (largest_component_ratio) of the total
-    points, an error is raised.
-    """
-    G = nx.Graph()
-    G.add_nodes_from(np.arange(stm_pnts.sizes["space"]))  # Use point indices as node identifiers
-    G.add_edges_from(
-        zip(
-            stm_arcs["source"].values,
-            stm_arcs["target"].values,
-            strict=True,
-        )
-    )
-
-    # Get list of connected components
-    list_components = [cc for cc in nx.connected_components(G)]
-
-    # If there are multiple connected components, keep only the largest one and discard the others
-    if len(list_components) > 1:
-        nodes_largest = max(list_components, key=len)  # set of node indices in the largest connected component
-
-        # Check if the largest component is significantly larger than the second largest one
-        if (len(nodes_largest) / stm_pnts.sizes["space"]) < largest_component_ratio:
-            raise RuntimeError(
-                f"The largest connected component contains only {len(nodes_largest)} points, which is less than "
-                f"{largest_component_ratio * 100:.1f}% of the total {stm_pnts.sizes['space']} points. "
-                "In this case DePSI cannot automatically decide which component to keep. "
-                "This may indicate a problem with the network formation. "
-                "Please check the input data and parameters."
-            )
-
-        mask_arcs = xr.DataArray(
-            np.isin(stm_arcs["source"].values, list(nodes_largest))
-            & np.isin(stm_arcs["target"].values, list(nodes_largest)),
-            dims=["space"],
-        )
-        stm_arcs_output = stm_arcs.where(mask_arcs, drop=True)
-        stm_pnts_output = stm_pnts.isel(space=list(nodes_largest))
-
-        # Update the source and target indices in stm_arcs_output to match the new stm_pnts_output
-        idx_map = {old_idx: new_idx for new_idx, old_idx in enumerate(sorted(nodes_largest))}
-        stm_arcs_output_updated = stm_arcs_output.copy()
-        stm_arcs_output_updated["source"] = xr.DataArray(
-            np.vectorize(idx_map.get)(stm_arcs_output["source"].values), dims="space"
-        )
-        stm_arcs_output_updated["target"] = xr.DataArray(
-            np.vectorize(idx_map.get)(stm_arcs_output["target"].values), dims="space"
-        )
-        stm_arcs_output = stm_arcs_output_updated
-
-        n_components = len(list_components)
-        logger.info("Separated components detected in the network!")
-        logger.info(f"Network has {n_components} connected components.")
-        logger.info(
-            f"Keeping only the largest component with {stm_pnts_output.sizes['space']} points "
-            f"and {stm_arcs_output.sizes['space']} arcs."
-        )
-        logger.info(
-            f"Discarded {stm_pnts.sizes['space'] - stm_pnts_output.sizes['space']} points "
-            f"and {stm_arcs.sizes['space'] - stm_arcs_output.sizes['space']} arcs."
-        )
-    else:
-        stm_arcs_output = stm_arcs
-        stm_pnts_output = stm_pnts
-
-    return stm_arcs_output, stm_pnts_output
-
-
-def _solve_float_ambiguities(A, y, invQy, sparse_mode: bool = False):
-    """Solve ambiguities as a float based on Least-Squares."""
-    # Solve ambiguities as they are float numbers
-    # This solves the equation y = Ax in least-squares sense
-    # With A a sparse matrix
-    # And stochastic model Qyy taken into account
-    invQyA = invQy @ A  # Avoid repeated computation in vectorized lsmr
-    invAtQyA = np.linalg.inv(A.T @ invQyA)
-
-    if sparse_mode:
-
-        @np.vectorize(signature="(i)->(j)")
-        def lsmr(y):
-            """Least-squares iterative solver for sparse data."""
-            x, *_ = scipy.sparse.linalg.lsmr(invQyA, invQy @ y)
-            return x
-
-        acheck = lsmr(y.T).T  # float ambiguity estimation
-    else:
-        acheck = invAtQyA @ (A.T @ invQy @ y)
-    echeck = y - A @ acheck  # residuals estimation
-
-    return acheck, echeck, invAtQyA
-
-
-def _remove_network_points_min_connections(
-    stm: xr.Dataset, arcs: xr.Dataset, min_connections: int
-) -> tuple[xr.Dataset, xr.Dataset]:
-    """Remove points which have less than min_connections arc connections.
-
-    The following steps are performed:
-
-    1. Remove points from stm which have less than min_connections connections in arcs.
-    2. Remove arcs which connect to the removed points.
-    3. Update the space indices in points/arcs STM accordingly.
-       The point indices are always a 0-based continuous array.
-
-    Note that this function does not perform iterative removal to assure that all points have
-    at least min_connections connections, but only performs one round of removal.
-    """
-    if min_connections < 1:
-        raise ValueError("min_connections must be at least 1")
-
-    # Load source and target indices from arcs
-    # these are 1d arrays so should fit in memory
-    idx_source = arcs["source"].values
-    idx_target = arcs["target"].values
-
-    # Select STM points that are in arcs
-    # Only keep points which ids are in arcs, isolated points are removed in idx_selected
-    idx_selected, counts = np.unique(np.concatenate([idx_source, idx_target]), return_counts=True)
-    idx_selected = idx_selected[counts >= min_connections]  # only keep points with at least min_connections connections
-
-    # If no change, return directly
-    if len(idx_selected) == stm.sizes["space"]:
-        return stm, arcs
-
-    # Select points
-    stm_pnts_updated = stm.isel(space=idx_selected)
-
-    # Select arcs that connect selected points (Some arcs may be dropped together with points)
-    mask_source = np.isin(idx_source, idx_selected)
-    mask_target = np.isin(idx_target, idx_selected)
-    mask_arcs = mask_source & mask_target
-    arcs = arcs.isel(space=np.where(mask_arcs)[0])
-
-    # The space size of the STM changes, resulting non-contiguous indices in space dimension
-    # hence an update in arcs space coordinates is needed
-    # Here we use a mapping solution, since the maximum number of network points is usually <100k
-    # Map old indices in arcs to new indices
-    idx_map = {old_idx: new_idx for new_idx, old_idx in enumerate(idx_selected)}
-    # apply the mapping to the source and target indices in arcs
-    arcs_updated = arcs.copy()
-    arcs_updated["source"] = xr.DataArray(np.vectorize(idx_map.get)(arcs["source"].values), dims="space")
-    arcs_updated["target"] = xr.DataArray(np.vectorize(idx_map.get)(arcs["target"].values), dims="space")
-
-    return stm_pnts_updated, arcs_updated
-
-
-def _generate_arcs_delaunay(coordinates, max_length):
-    """Create a network using Delaunay triangulation."""
-    # Create network and collect neighbors.
-    network = Delaunay(coordinates)
-    neighbors_ptr, neighbors_idx = network.vertex_neighbor_vertices
-
-    # Convert ptr and idx arrays into list of sorted index pairs.
-    arcs = []
-    for s in range(len(neighbors_ptr) - 1):
-        for t in range(neighbors_ptr[s], neighbors_ptr[s + 1]):
-            length = get_distance(coordinates[int(s)], coordinates[neighbors_idx[t]], mode="euclidean")
-            if max_length is None or length <= max_length:
-                arcs.append(tuple(sorted([int(s), int(neighbors_idx[t])])))
-
-    # Remove duplicates and make the list canonical.
-    arcs = sorted(list(set(arcs)))
-
-    return arcs
-
-
-def _generate_arcs_redundant(coordinates, max_length, min_links, num_partitions):
-    """Create a network with at least min_links arcs per node.
-
-    The redundant network is formed with the following steps:
-
-    1. Create a KDTree and find all pairs of points within the maximum distance.
-    2. Loop through each point and find its neighbors within the maximum distance.
-    3. Divide neighbors into partitions based on their direction.
-    4. Select the nth nearest neighbors from all partitions, starting from n=1.
-    5. Sort the selected neighbors by distance, add them to the arcs list. If min_links is not
-       exceeded, continue to the n+1th nearest neighbors of all partitions.
-    6. Repeat until min_links is reached.
-    """
-    arcs = []
-    indices = range(len(coordinates))
-
-    # Create a KDTree for fast distance queries.
-    tree = KDTree(coordinates)
-    if max_length is not None:
-        pairs = tree.query_pairs(r=max_length, output_type="ndarray")
-    else:
-        pairs = tree.query_pairs(r=np.inf, output_type="ndarray")
-
-    # Duplicate pairs with reversed indices to ensure that arcs are undirected.
-    pairs = np.concatenate((pairs, np.flip(pairs, axis=1)), axis=0)
-    pairs = pairs[np.argsort(pairs[:, 0])]  # Sort pairs by first column (source index).
-
-    for cur_index in indices:
-        # Get the neighbors of the current node
-        neighbors = pairs[pairs[:, 0] == cur_index][:, 1].tolist()
-
-        if len(neighbors) == 0:  # skip if there are no neighbors
-            continue
-        elif len(neighbors) <= min_links:
-            # If there are not enough neighbors, connect them all.
-            for idx in neighbors:
-                arc_to_add = (min(cur_index, idx), max(cur_index, idx))
-                arcs.append(arc_to_add)
-        else:
-            # Calculate partitions and distances for neighbors
-            partitions = [
-                int(math.floor(num_partitions * (0.5 + math.atan2(coordinate[1], coordinate[0]) / math.tau)))
-                for coordinate in coordinates[neighbors] - coordinates[cur_index]
-            ]
-            distances = [math.dist(coordinates[cur_index], coordinates[idx]) for idx in neighbors]
-
-            # Create sorted array by partition and then distance
-            sorted_arr = np.array(sorted(list(zip(partitions, distances, neighbors, strict=False))))
-
-            # Split into partitions
-            partitions_diff = sorted_arr[1:, 0] - sorted_arr[:-1, 0]
-            separators = np.where(partitions_diff > 0)[0]
-            partitions_split = np.split(sorted_arr, separators + 1)
-            partitions_split = [partition[:min_links] for partition in partitions_split]
-
-            # Collect the neighbor 'hierarchies'
-            neighbor_hierarchies = [[] for _ in range(min_links)]
-            count = 0
-            for n in range(min_links):
-                # Break early if we have gathered enough neighbors.
-                if min_links <= count:
-                    break
-                for partition in partitions_split:
-                    # Note that we do not break inside this loop,
-                    # because we want the nth nearest neighbors from all partitions.
-                    if n < len(partition) and (max_length is None or partition[n][1] <= max_length):
-                        neighbor_hierarchies[n].append(partition[n])
-                        count = count + 1
-
-            # Sort hierarchies per partition by distance to the current node
-            neighbor_hierarchies = [
-                sorted(hierarchy, key=lambda x: x[1]) for hierarchy in neighbor_hierarchies if len(hierarchy) != 0
-            ]
-
-            # Add sorted arcs to at least min_links neighbors
-            cur_arcs = [
-                (min(cur_index, int(neighbor[2])), max(cur_index, int(neighbor[2])))
-                for hierarchy in neighbor_hierarchies
-                for neighbor in hierarchy
-            ]
-            cur_arcs = cur_arcs[:min_links]
-
-            arcs.extend(cur_arcs)
-
-    # Remove duplicates and make the list canonical.
-    arcs = sorted(list(set(arcs)))
-
-    return arcs
-
-
-def get_ordered_arcs(
-    x_ref_search,
-    y_ref_search,
-    buffer_radius_ref,
-    dist_to_quality,
-    n_max_arcs,
-    nad_nmad_max,
-    x_coordinates,
-    y_coordinates,
-    slc_quality,
-    nad_nmad_vals,
-    coordinate_type,
-):
-    """Get a list with ordered arcs based on pnt quality and a search area.
-
-    Args:
-    ----
-        x_ref_search (float): x-coordinate of the centre of the search area
-        y_ref_search (float): y-coordinate of the centre of the search area
-        buffer_radius_ref (float): The buffer (in m) around the centre coordinates where potential arcs are computed
-        dist_to_quality (float): parameter that relates arc length to additional sigma
-        n_max_arcs (float): The maximum nr of arcs to be outputed
-        nad_nmad_max (float): The maximum NAD / NMAD, for the entire time series, for a point to be considered
-        x_coordinates (xr.DataArray): x-coordinates of all points in the STM
-        y_coordinates (xr.DataArray): y-coordinates of all points in the STM
-        slc_quality (xr.DataArray): quality of the estimates per partition
-        nad_nmad_vals (xr.DataArray): NAD / NMAD of the entire time series per point
-        coordinate_type: 'Euclidean' (RD) / 'geographic' (lon/lat)
-
-    Returns:
-    -------
-        arcs (list): ordered arcs
-        arcs_and_quality (list): ordered list with the points and quality
-        quality_dict (dictionary): dictionary with the arcs and their quality
-    """
-    # Find all points within the buffer around the starting location x, y
-    idx_pnts_buffer_ref = find_points_within_buffer(
-        x_coordinates, y_coordinates, x_ref_search, y_ref_search, buffer_radius_ref, coordinate_type
-    )
-
-    # Get the a-priori quality of all potential arcs that can be made
-    arcs_and_quality = _ordered_arcs_all_points(
-        x_coordinates,
-        y_coordinates,
-        slc_quality,
-        idx_pnts_buffer_ref,
-        dist_to_quality,
-        nad_nmad_vals,
-        nad_max=nad_nmad_max,
-    )
-
-    # We will only work with n_max_arcs, otherwise we need to load an extensive dataset everytime
-    arcs_and_quality = arcs_and_quality[0:n_max_arcs]
-
-    # Only store the arcs, and remove the quality values
-    arcs = []
-    for _, arc in arcs_and_quality:
-        arcs.append(arc)
-
-    # Create a dictionary with the quality of the arcs
-    quality_dict = {tuple(sorted(arc[1])): arc[0] for arc in arcs_and_quality}
-
-    return arcs, arcs_and_quality, quality_dict
-
-
-def find_points_within_buffer(
-    x_coords,
-    y_coords,
-    x_pnts,
-    y_pnts,
-    buffer_radius,
-    coordinate_type: Literal["euclidean", "geographic"] = "euclidean",
-    return_aggregate_point_buffer: bool = True,
-):
-    """Find all points located within a specified buffer radius around a given location.
-
-    This function determines which points in a set of coordinates are located within a defined buffer
-    distance around one or more specified points.
-
-    The function:
-    1. Converts the input x and y coordinates of the search point(s) to arrays if they are not already.
-    2. Computes the distance between each point in `x_coords` and `y_coords` and the reference points.
-    3. Returns the indices of these points for further processing or analysis.
-
-    Args:
-    ----
-        x_coords (numpy.ndarray): Array of x-coordinates for all points in the dataset.
-        y_coords (numpy.ndarray): Array of y-coordinates for all points in the dataset.
-        x_pnts (float or array-like): The x-coordinate or array of x-coordinates of the reference point(s).
-        y_pnts (float or array-like): The y-coordinate or array of y-coordinates of the reference point(s).
-        buffer_radius (float): The radius of the buffer zone around the reference points, specified in meters.
-        coordinate_type (Literal["euclidean", "geographic"]): whether the coordinates provided are Euclidean (such as
-            RD) or geographic (such as lon / lat)
-        return_aggregate_point_buffer (bool): if True, a list of indices of the points within the buffer of any of the
-            points in x_pnts/y_pnts is returned. If False, the list of indices within the buffer of each point in
-            x_pnts/y_pnts is returned individually. Default True
-
-    Returns:
-    -------
-        indices (numpy.ndarray): Array of indices of the points located within the buffer zone, either the aggregate
-            buffer zone (`return_aggregate_point_buffer`=True) or per point in `x_pnts`/`y_pnts`
-             (`return_aggregate_point_buffer`=False)
-
-    Example:
-        indices = find_points_within_buffer(x_coords, y_coords, x_pnts=10.5, y_pnts=20.3, buffer_radius=5.0)
-        (with Euclidean coordinates and an aggregate buffer output)
-    """
-    # Make sure x_pnts and y_pnts are arrays
-    x_pnts = np.atleast_1d(x_pnts)
-    y_pnts = np.atleast_1d(y_pnts)
-    assert len(x_pnts.shape) == 1, f"Received more than 1 dimension in x_pnts (shape {x_pnts.shape})!"
-    assert len(y_pnts.shape) == 1, f"Received more than 1 dimension in y_pnts (shape {y_pnts.shape})!"
-
-    if coordinate_type == "euclidean":
-        # could possibly also be a KDTree but for consistency inside the function the minkowski metric does Euclidean
-        tree = BallTree(np.vstack([x_coords, y_coords]).T, metric="minkowski")
-        search_radius = buffer_radius
-        search_points = np.vstack([x_pnts, y_pnts]).T
-    elif coordinate_type == "geographic":
-        # the geographic tree assumes an Earth radius of 1, so we need to divide the search radius by the radius of
-        # the Earth
-        # It also expects input in radians instead of degrees, and first latitude (y), then longitude (x)
-        tree = BallTree(np.vstack([np.radians(y_coords), np.radians(x_coords)]).T, metric="haversine")
-        search_radius = buffer_radius / EARTH_RADIUS
-        search_points = np.vstack([np.radians(y_pnts), np.radians(x_pnts)]).T
-    else:
-        raise ValueError(f"Unknown coordinate type {coordinate_type}! Known are euclidean and geographic.")
-
-    res_indices = tree.query_radius(search_points, search_radius)
-
-    if not return_aggregate_point_buffer:
-        return res_indices
-
-    all_indices = []
-    for i in res_indices:  # flatten won't work since res_indices is an irregularly shaped numpy array
-        for j in i:
-            all_indices.append(j)
-    indices = np.array(list(sorted(list(set(all_indices)))))
-
-    return indices
-
-
-def _ordered_arcs_all_points(rdx, rdy, slc_quality, idx_pnts_buffer, dist_to_quality, nad_vals, nad_max):
-    """Generate a sorted list of unique arcs between all points within buffer, ranked by combined quality and distance.
-
-    This function computes arcs between all points in a given buffer based on their spatial distance and
-    quality metrics. The arcs are ranked according to a combination of the Euclidean distance between
-    points and the maximum quality value between them.
-
-    The function:
-    1. Computes the Euclidean distance matrix for all points within the buffer.
-    2. Calculates the arc quality time series for all point pairs based on their SLC quality values.
-    3. Extracts the maximum quality for each arc and combines it with the distance between the points.
-    4. Uses the lower triangular matrix to avoid duplicate arcs
-    5. Sorts the arcs based on the combined distance and quality metric.
-
-    Args:
-    ----
-        rdx (numpy.ndarray): Array of x-coordinates for all points.
-        rdy (numpy.ndarray): Array of y-coordinates for all points.
-        slc_quality (numpy.ndarray): Array of SLC quality time series for all points.
-        idx_pnts_buffer (numpy.ndarray): Indices of the points within the buffer.
-        dist_to_quality (float): Scaling factor for weighting the distance in combination with the quality.
-        nad_vals (numpy.ndarray): Array of NAD values of all points.
-        nad_max (float): The maximum NAD, for the entire time series, for a point to be considered
-
-    Returns:
-    -------
-        arcs_and_quality (list): Sorted list of tuples, each tuple contains the combined quality and distance value
-                                 and the indices of the two points forming the arc.
-
-    Example:
-    -------
-        arcs_and_quality = ordered_arcs_all_points(rdx, rdy, slc_quality, idx_pnts_buffer, dist_to_quality)
-    """
-    # Compute NAD values of the points within the buffer
-    nad_buffer_vals = nad_vals[idx_pnts_buffer]
-
-    # Mask points that have an NAD value above the threshold
-    mask_nad = nad_buffer_vals < nad_max
-    idx_pnts_buffer = idx_pnts_buffer[mask_nad]
-
-    # Compute the distance matrix for all points (below the NAD threshold) within the buffer
-    rdx_buffer = rdx[idx_pnts_buffer]
-    rdy_buffer = rdy[idx_pnts_buffer]
-    coords_points = np.vstack((rdx_buffer, rdy_buffer)).T
-    dist_matrix = distance_matrix(coords_points, coords_points)
-
-    # Compute the quality matrix for all potential arcs within the buffer
-    slc_quality_buffer = slc_quality[idx_pnts_buffer, :]
-    slc_quality_i_buffer = slc_quality_buffer[:, np.newaxis]
-    slc_quality_j_buffer = slc_quality_buffer[np.newaxis, :]
-
-    # Compute the arc quality time series
-    # quality is based on both sides of the arc
-    arc_quality_ts = np.sqrt(slc_quality_i_buffer**2 + slc_quality_j_buffer**2)
-
-    # Get the maximum value in the time dimension per arc
-    # This is considered as the 'worst' quality for the entire time period
-    arcs_quality_max = np.max(arc_quality_ts, axis=2)
-
-    # Get only the lower triangular matrix
-    arcs_quality_max_lower = np.tril(arcs_quality_max)
-    arcs_dist = np.tril(dist_matrix)
-
-    # Add additional sigma because of the arc length
-    arc_quality_dist_max = arcs_quality_max_lower + arcs_dist * dist_to_quality
-
-    # Order the arcs
-    arcs_and_quality = []
-
-    # Loop trough the lower tringular matrix and add values list
-    rows, cols = arc_quality_dist_max.shape
-    for i in range(rows):
-        for j in range(i):  # j < i makes sure that we only get the lower triangular matrix
-            arcs_and_quality.append((arc_quality_dist_max[i, j], (idx_pnts_buffer[i], idx_pnts_buffer[j])))
-
-    # Sort the list with all the arcs from best to worse
-    arcs_and_quality.sort()
-
-    return arcs_and_quality
-
-
-def construct_control_network(
-    arcs,
-    quality_dict_arcs,
-    excluded_arcs,
-    n_top,
-    n_batch,
-    deg_threshold,
-    min_nodes,
-    min_redundancy,
-    visualize_network=False,
-):
-    """Construct a control network of arcs based on quality metrics and structural requirements.
-
-    This function iteratively builds a network from a search area of arcs, filtering out excluded arcs
-    (where an excluded arc is an arc we know we don't want to have) and ensuring that the resulting
-    network meets specified criteria, such as a minimum number of nodes and redundancy.
-    The network is constructed by:
-    1. Ranking arcs based on quality metrics provided in `quality_dict_arcs`.
-    2. Excluding arcs that are in the `excluded_arcs` list.
-    3. Iteratively adding arcs in batches and refining the network to remove nodes with low centrality.
-    4. Evaluating the network against requirements such as minimum nodes and average redundancy.
-
-    If the requirements are not met with the given arcs, a warning is issued. Optionally, the network
-    construction process can be visualized.
-
-    Args:
-    ----
-        arcs (list of tuples): List of arcs (pairs of points) within the search area.
-        quality_dict_arcs (dict): Dictionary mapping arcs (tuples) to a sigma value.
-        excluded_arcs (list of tuples): List of arcs that should be excluded from the network.
-        n_top (int): Number of top-ranked arcs to start with.
-        n_batch (int): Number of additional arcs to add in each iteration.
-        deg_threshold (float): Degree threshold for removing low-centrality nodes from the network.
-        min_nodes (int): Minimum number of nodes required in the final network.
-        min_redundancy (float): Minimum average redundancy (degree) required in the final network.
-        visualize_network (bolean, optional): If set to True, the network construction process is visualized.
-
-    Returns:
-    -------
-        ref_pnt (int): The reference point used in the final iteration of the network.
-        arcs_updated_network_sorted (list of tuples): The sorted list of arcs in the final network
-            based on quality metrics.
-        ref_pnt_initial (int): The reference point used in the initial network.
-        arcs_initial_network (list of tuples): The list of arcs in the network before refinement.
-
-    Example:
-    -------
-        ref_pnt, final_arcs, initial_ref_pnt, initial_arcs = construct_control_network(
-            arcs, quality_dict_arcs, excluded_arcs,
-            n_top=50, n_batch=10,
-            deg_threshold=1.5, min_nodes=20, min_redundancy=2.0,
-            visualize_network=False
-        )
-    """
-    # Sort arcs by quality, removing excluded arcs from the ranked arcs variable
-    excluded_arcs_sorted = [tuple(sorted(arc)) for arc in excluded_arcs]
-    arcs_without_excluded = [arc for arc in arcs if tuple(sorted(arc)) not in excluded_arcs_sorted]
-
-    # Initialize variables
-    current_network = None
-    avg_degree = 0
-    iteration = 0
-
-    more_arcs_to_test = True
-    network_reqs_not_met = True
-
-    # the following while loop will stop when either the network requirements (# nodes and redundancy level) are met,
-    # OR when we run out of arcs to text
-    while more_arcs_to_test and network_reqs_not_met:
-        # Select the arcs for the current iteration
-        iteration += 1
-        end_idx = n_top + iteration * n_batch
-        arcs_to_test = arcs_without_excluded[0:end_idx]
-
-        if not arcs_to_test:  # Break if no more arcs to add
-            print("No more arcs to test.")
-            more_arcs_to_test = False
-            continue
-
-        # Create a new network using the selected arcs
-        current_network = nx.Graph()
-        current_network, _, ref_pnt = _from_arcs_to_graph(arcs_without_excluded[:end_idx], plot=visualize_network)
-        if iteration == 1:
-            ref_pnt_initial = ref_pnt
-            arcs_initial_network = [tuple(sorted(arc)) for arc in current_network.edges()]
-
-        # Remove nodes with degree equal or lower than the threshold
-        current_network, _, ref_pnt = _remove_low_centrality_nodes(
-            current_network, deg_threshold=deg_threshold, plot=visualize_network
-        )
-        arcs_updated_network = [tuple(sorted(arc)) for arc in current_network.edges()]
-
-        # Test the network against requirements
-        avg_degree = len(current_network.edges) / len(current_network.nodes) if len(current_network.nodes) > 0 else 0
-        print(f"The average degree is {avg_degree:.2f}")
-
-        if len(current_network.nodes) >= min_nodes and avg_degree >= min_redundancy:
-            network_reqs_not_met = False
-            # Stop if requirements are met
-
-    # Final check if the network meets requirements
-    if len(current_network.nodes) < min_nodes or avg_degree < min_redundancy:
-        print("Warning: Network could not meet all requirements with the given arcs.")
-
-    # The arcs constructed above are no longer sorted based on their quality
-    arcs_updated_network_sorted = sorted(
-        arcs_updated_network, key=lambda arc: quality_dict_arcs.get(tuple(sorted(arc)), float("inf"))
-    )
-
-    return ref_pnt, arcs_updated_network_sorted, ref_pnt_initial, arcs_initial_network
-
-
-def _from_arcs_to_graph(arcs, plot=False, save_path="./network.png"):
-    """Construct a graph from a set of arcs and identifies key properties of the network.
-
-    Args:
-    ----
-        arcs : (list of tuple)
-            List of arcs (edges) in the graph. Each arc is represented as a tuple
-            of two nodes, e.g., [(node1, node2), (node2, node3)].
-        plot : (boolean)
-            Default is False, If True, the function visualizes the graph structure, including highlighting
-            connected components. Default is False (no visualization).
-        save_path : (string)
-            Default = "./network.png"
-
-    Returns:
-    -------
-        tuple: A tuple containing:
-            - network (networkx.Graph): The constructed graph based on the provided arcs.
-            - degree_centrality (dict): A dictionary where keys are nodes and values are their
-                                        degree centrality scores.
-            - ref_pnt (int): The index of the reference point (point with most connections).
-    """
-    if not arcs:
-        raise ValueError("The input arcs list is empty. Please provide a valid list of arcs.")
-
-    # Compute the network
-    network = nx.Graph()
-    network.add_edges_from(arcs)
-
-    # Compute the degree centrality
-    # Which says something on how connected one particular node is with other nodes
-    degree_centrality = nx.degree_centrality(network)
-
-    # Get the reference point (node with the most connections)
-    ref_pnt = max(degree_centrality, key=degree_centrality.get)
-
-    if plot:
-        # Visualize the network
-        pos = nx.spring_layout(network)  # Lay-out for the graph
-        nx.draw(
-            network,
-            pos,
-            with_labels=True,
-            node_size=500,
-            node_color="skyblue",
-            font_size=10,
-            font_weight="bold",
-            edge_color="gray",
-        )
-
-        connected_components = list(nx.connected_components(network))
-        # Highlight the connected components
-        for component in connected_components:
-            nx.draw_networkx_nodes(network, pos, nodelist=component, node_color="orange", node_size=700)
-
-        # Highlight the reference point
-        nx.draw_networkx_nodes(network, pos, nodelist=[ref_pnt], node_color="red", node_size=800)
-        plt.title(f"Network Visualization - {len(connected_components)} Connected Components")
-        plt.savefig(save_path, bbox_inches="tight")
-
-    return network, degree_centrality, ref_pnt
-
-
-def _remove_low_centrality_nodes(network, deg_threshold, plot=False, save_path="./network.png"):
-    """Remove nodes with a low degree of centrality.
-
-    Args:
-    ----
-        network (networkx.Graph): The input graph from which nodes will be removed.
-        deg_threshold (int): The degree threshold; nodes with a degree equal to or less than this value will be removed.
-        plot (bolean, optional): If True, the function visualizes the updated graph, including its connected components.
-                              Default is False (no visualization).
-        save_path : (string)
-            Default = "./network.png"
-
-    Returns:
-    -------
-        tuple: A tuple containing:
-            - network (networkx.Graph): The updated graph after removing nodes with the specified degree threshold.
-            - degree_centrality (dict): A dictionary where keys are nodes and values are their degree centrality scores.
-            - ref_pnt (int or None): The node with the highest degree centrality in the updated graph.
-
-    Example:
-    -------
-        updated_network, degree_centrality, ref_point = remove_low_centrality_nodes(
-            network, deg_threshold=1, plot=False
-        )
-    """
-    # Identify nodes to remove
-    nodes_to_remove = [node for node, degree in network.degree() if degree <= deg_threshold]
-    removed_nodes = set(nodes_to_remove)  # For plot highlighting
-
-    # Remove these nodes from the network
-    network.remove_nodes_from(nodes_to_remove)
-
-    # Handle empty graph case
-    if not network.nodes:
-        return network, {}, None
-
-    # Degree centrality
-    degree_centrality = nx.degree_centrality(network)
-
-    # Get the reference point (node with the highest degree centrality)
-    ref_pnt = max(degree_centrality, key=degree_centrality.get)
-
-    if plot:
-        # Visualize the updated network
-        pos = nx.spring_layout(network)
-        nx.draw(
-            network,
-            pos,
-            with_labels=True,
-            node_size=500,
-            node_color="skyblue",
-            font_size=10,
-            font_weight="bold",
-            edge_color="gray",
-        )
-
-        connected_components = list(nx.connected_components(network))
-
-        # Highlight connected components
-        for component in connected_components:
-            nx.draw_networkx_nodes(network, pos, nodelist=component, node_color="orange", node_size=700)
-
-        # Highlight removed nodes in the plot (optional)
-        if removed_nodes:
-            # Filter valid nodes that exist in the layout
-            valid_removed_nodes = [node for node in removed_nodes if node in pos]
-            nx.draw_networkx_nodes(network, pos, nodelist=valid_removed_nodes, node_color="red", node_size=800)
-
-        nx.draw_networkx_nodes(network, pos, nodelist=[ref_pnt], node_color="red", node_size=800)
-
-        plt.title(f"Network Visualization after Removing Nodes with Degree ≤ {deg_threshold}")
-        plt.savefig(save_path, bbox_inches="tight")
-
-    return network, degree_centrality, ref_pnt
-
-
-def test_succeeded_arcs_control_network(
-    succeeded_arcs, quality_dict_arcs, deg_threshold, min_nodes, min_redundancy, visualize_network=False
-):
-    """Evaluate a network constructed from a set of arcs to determine if it meets structural requirements.
-
-    This function tests a network built from a set of arcs to ensure it satisfies minimum structural criteria,
-    including the number of nodes, redundancy, and centrality. It checks whether any nodes with low degree should
-    be removed and then evaluates the network again. The arcs are sorted by quality after refinement.
-
-    Parameters
-    ----------
-    succeeded_arcs : list of tuples
-        A list of arcs (pairs of points) that are used to construct the network.
-    quality_dict_arcs : dict
-        A dictionary where the keys are the arcs and the values are their respective quality scores.
-    deg_threshold : float
-        The degree threshold for identifying low-centrality nodes. Nodes with degrees smaller will be removed.
-    min_nodes : int
-        The minimum number of nodes required for the network to be valid.
-    min_redundancy : float
-        The minimum average degree (redundancy) required for the network.
-    visualize_network : bool, optional
-        A flag to control network visualization of the network during evaluation. Default is False (no visualization).
-
-    Returns
-    -------
-    network_check : int
-        A flag indicating whether the network meets the requirements (1 if successful, 0 if not).
-    arcs_updated_network_sorted : list of tuples
-        The sorted list of arcs based on their quality after refining the network.
-    ref_pnt : int
-        The reference point used in the network.
-
-    Example
-    -------
-    network_check, final_arcs, ref_pnt = test_succeeded_arcs_control_network(
-        succeeded_arcs, deg_threshold=1.5, min_nodes=20,
-        min_redundancy=2.0, visualize_network=1
-    )
-    """
-    arcs_updated_network_sorted = []
-    network_check = 0
-
-    # Construct the network based on the input arcs
-    current_network = nx.Graph()
-    current_network, _, ref_pnt = _from_arcs_to_graph(succeeded_arcs, plot=visualize_network)
-    arcs_updated_network = [tuple(sorted(arc)) for arc in current_network.edges()]
-
-    # Check whether there are nodes with degree smaller than one
-    nodes_to_remove = [node for node, degree in current_network.degree() if degree <= deg_threshold]
-
-    if not nodes_to_remove:
-        # Test the network against requirements
-        avg_degree = len(current_network.edges) / len(current_network.nodes) if len(current_network.nodes) > 0 else 0
-        print(f"The average degree is {avg_degree:.2f}")
-
-        if len(current_network.nodes) >= min_nodes and avg_degree >= min_redundancy:
-            network_check = 1
-
-            # The arcs constructed above are no longer sorted based on their quality
-            arcs_updated_network_sorted = sorted(
-                arcs_updated_network, key=lambda arc: quality_dict_arcs.get(tuple(sorted(arc)), float("inf"))
-            )
-
-    else:
-        print("We remove low degree nodes")
-        # Remove the nodes with low degree and test the network again
-        current_network, _, ref_pnt = _remove_low_centrality_nodes(
-            current_network, deg_threshold=deg_threshold, plot=visualize_network
-        )
-        arcs_updated_network = [tuple(sorted(arc)) for arc in current_network.edges()]
-
-        # Test the network against requirements
-        avg_degree = len(current_network.edges) / len(current_network.nodes) if len(current_network.nodes) > 0 else 0
-        print(f"The average degree is {avg_degree:.2f}")
-
-        if len(current_network.nodes) >= min_nodes and avg_degree >= min_redundancy:
-            network_check = 1
-
-            # The arcs constructed above are no longer sorted based on their quality
-            arcs_updated_network_sorted = sorted(
-                arcs_updated_network, key=lambda arc: quality_dict_arcs.get(tuple(sorted(arc)), float("inf"))
-            )
-
-        else:
-            network_check = 0
-
-    return network_check, arcs_updated_network_sorted, ref_pnt
-
-
-def ordered_arcs_connection_point_and_control_network(
-    rdx_connection_point,
-    rdy_connection_point,
-    slc_quality_connection_point,
-    connection_point_idx,
-    rdx_control,
-    rdy_control,
-    slc_quality_control,
-    control_idx,
-    dist_to_quality,
-    coordinate_type: Literal["euclidean", "geometric"],
-):
-    """Generate a sorted array of unique arcs between a 'connection_point' and the control network.
-
-    This function computes arcs between a connection point and control points, evaluating each arc based on
-    a combination of spatial distance and SLC quality. The arcs are sorted from best to worst, where "best"
-    is determined by the lowest combined quality and distance value.
-
-    The function:
-    1. Computes the Euclidean distance between the connection point and the control points.
-    2. Calculates the arc quality time series for each arc.
-    3. Combines the distance and maximum quality for each arc to define an overall quality value.
-    4. Sorts the arcs by quality values from best (lowest) to worst (highest).
-    5. Constructs a sorted array of arcs, ensuring the control point appears first in each arc.
-
-    Args:
-    ----
-        rdx_connection_point (xarray.DataArray): x-coordinate of the connection point.
-        rdy_connection_point (xarray.DataArray): y-coordinate of the connection point.
-        slc_quality_connection_point (xarray.DataArray): Array of SLC quality time series for the connection point.
-        connection_point_idx (int): Index of the connection point.
-        rdx_control (xarray.DataArray): x-coordinates of the control points.
-        rdy_control (xarray.DataArray): y-coordinates of the control points.
-        slc_quality_control (xarray.DataArray): Array of SLC quality time series for the control points.
-        control_idx (xarray.DataArray): Indices of the control points.
-        dist_to_quality (float): Conversion factor to scale distance relative to quality.
-        coordinate_type (Literal["euclidean", "geographic"]): whether the provided coordinates are Euclidean or
-            geographic
-
-    Returns:
-    -------
-        tuple:
-            - arcs (numpy.ndarray): A sorted array of arcs, where each row contains a control point
-              followed by the connection point.
-            - sorted_quality_values (numpy.ndarray): Sorted quality values corresponding to the arcs.
-    """
-    # Stack the coordinates of the connection point and the control points
-    coords_connection_point = np.vstack((rdx_connection_point, rdy_connection_point)).T
-    coords_control = np.vstack((rdx_control, rdy_control)).T
-
-    # Distance matrix between the connection point and the control points
-    if coordinate_type == "euclidean":
-        dist_matrix = distance_matrix(coords_connection_point, coords_control).squeeze()
-    else:
-        dist_matrix = np.zeros((coords_connection_point.shape[0], coords_control.shape[0]))
-        for m in range(coords_connection_point.shape[0]):
-            for n in range(coords_control.shape[0]):
-                dist_matrix[m, n] = get_distance(coords_connection_point[m], coords_control[n], mode=coordinate_type)
-
-    # # Compute the quality matrix for connection point (pnt i) and all control points (point j)
-    slc_quality_j = slc_quality_control.values  # Quality values of the control points
-    slc_quality_i = np.expand_dims(slc_quality_connection_point.values, axis=0)  # Make sure the dimensions match
-    slc_quality_i = np.repeat(slc_quality_i, repeats=slc_quality_j.shape[0], axis=0)
-
-    # Compute arc quality time series (which is a function of the slc_quality of point i and point j)
-    arc_quality_ts = np.sqrt(slc_quality_i**2 + slc_quality_j**2)
-    arc_quality_max = np.max(arc_quality_ts, axis=1)  # Compute the maximum value
-
-    # Combine distance and quality
-    arc_quality_dist_max = arc_quality_max + dist_matrix * dist_to_quality
-
-    # Sort the arcs and compute the indices of the control points (since point i, the connection_point, is in all arcs)
-    sorted_control = np.argsort(arc_quality_dist_max)
-    sorted_control_idx = control_idx.values[sorted_control]  # sort the indices of the control points as well
-    sorted_quality_values = np.sort(arc_quality_dist_max)
-
-    # Compute the arcs between the connection_point and the control points.
-    # Make sure that the control_points comes first
-    arcs = np.zeros((len(sorted_control_idx), 2), dtype=int)
-    arcs[:, 0] = sorted_control_idx  # control network points
-    arcs[:, 1] = connection_point_idx  # connection_point
-
-    return arcs, sorted_quality_values
-
-
-def construct_control_network_test_arcs(
-    x_ref_search,
-    y_ref_search,
-    buffer_radius_ref,
-    dist_to_quality,
-    N_max_arcs,
-    N_top,
-    N_batch,
-    deg_threshold,
-    min_nodes,
-    min_redundancy,
-    nad_nmad_max,
-    visualize_network,
-    sigma_post_over_sigma_prior,
-    nr_max_iter_control,
-    bounds,
-    m2ph,
-    years,
-    dates,
-    temperature,
-    sd_complex,
-    slc_quality,
-    cr2ph,
-    ampl_ts,
-    bkps_stm,
-    mean_ampl_sd,
-    sigma_ampl_sd,
-    mad_ampl_sd,
-    median_ampl_sd,
-    x_coordinates,
-    y_coordinates,
-    nad_nmad,
-    coordinate_type: Literal["euclidean", "geographic"] = "euclidean",
-):
-    """Construct the control network while testing the arcs inside.
-
-    Parameters
-    ----------
-    x_ref_search:
-        X coordinate of central point to search for reference point
-    y_ref_search:
-        Y coordinate of central point to search for reference point
-    buffer_radius_ref:
-        Size of search window for reference point
-    dist_to_quality:
-        relates arc length to additional sigma
-    N_max_arcs:
-        Maximum number of arcs
-    N_top:
-        Number of top-ranked arcs to start with.
-    N_batch:
-        Number of additional arcs to add in each iteration.
-    deg_threshold:
-        Degree threshold for removing low-centrality nodes from the network.
-    min_nodes:
-        Minimum number of nodes required in the final network.
-    min_redundancy:
-        Minimum average redundancy (degree) required in the final network.
-    nad_nmad_max:
-        Maximum NAD / NMAD for a point to be considered
-    visualize_network:
-        Boolean whether or not to visualize the network
-    sigma_post_over_sigma_prior:
-        Upper limit on how much the aposteriori sigma of an arc is allowed to differ from the apriori sigma
-    nr_max_iter_control:
-        Maximum number of iterations in the network testing
-    bounds: list of tuples
-        Bounds for parameter estimation in the format (lower_bounds, upper_bounds).
-    m2ph: float
-        Meters to phase conversion factor
-    years: np.ndarray
-        Array of decimal years corresponding to the time series epochs.
-    dates: np.ndarray
-        Array of date indices or timestamps corresponding to the time series.
-    temperature: np.ndarray
-        Array of temperature values for thermal expansion modeling.
-    sd_complex: np.ndarray
-        Single-difference complex
-    slc_quality: np.ndarray
-        Complex-valued standard deviations of the signal for all points.
-    cr2ph: np.ndarray
-        crossrange-to-phase for each point
-    ampl_ts: np.ndarray
-        amplitude timeseries
-    bkps_stm: np.ndarray
-        Breakpoints stm
-    mean_ampl_sd: np.ndarray
-        Single difference mean amplitude
-    sigma_ampl_sd: np.ndarray
-        Single difference mean amplitude standard deviation
-    mad_ampl_sd: np.ndarray
-        Median absolute deviation of the single difference amplitude
-    median_ampl_sd: np.ndarray
-        Median single difference amplitude
-    x_coordinates: np.ndarray
-        x coordinates (RD or longitude)
-    y_coordinates: np.ndarray
-        y coordinates (RD or latitude)
-    nad_nmad: np.ndarray
-        NAD or NMAD to be used in the point selection upon which `nad_nmad_max` is imposed.
-    coordinate_type: Literal["euclidean", "geographic"] = "euclidean"
-        Whether the provided coordinates are Euclidean (such as RD) or geographic (such as lon/lat)
-
-    Returns
-    -------
-    dict
-        Results of the accepted network
-    list
-        Reference point
-    list
-        Results of the accepted arcs
-    """
-    # At the start, no arcs are tested yet, so we don't have failed_arcs, and noisy_arcs
-    failed_arcs = []
-    noisy_arcs = []
-
-    # Get an ordered list of all potential arcs with a buffer area
-    arcs_search_area, _, quality_dict_arcs = get_ordered_arcs(
-        x_ref_search,
-        y_ref_search,
-        buffer_radius_ref,
-        dist_to_quality,
-        N_max_arcs,
-        nad_nmad_max,
-        x_coordinates,
-        y_coordinates,
-        slc_quality,
-        nad_nmad,
-        coordinate_type,
-    )
-
-    network_meets_requirements = False
-
-    while not network_meets_requirements:
-        # Create a first network based on the ranked quality and leaving out the failed_arcs and noisy_arcs
-        ref_pnt, arcs_updated_network, ref_pnt_initial, arcs_initial_network = construct_control_network(
-            arcs_search_area,
-            quality_dict_arcs,
-            failed_arcs,
-            N_top,
-            N_batch,
-            deg_threshold,
-            min_nodes,
-            min_redundancy,
-            False,
-        )
-
-        print("Computing the solutions for the arcs")
-        # Test whether solutions for all arcs can be found (sometimes it happens that because of the complex
-        # functions no solutions can be found)
-        results_initial_control_network_v1 = arc_estimation_control_network(
-            arcs_updated_network,
-            bounds,
-            m2ph,
-            nr_max_iter_control,
-            years,
-            dates,
-            temperature,
-            sd_complex,
-            slc_quality,
-            cr2ph,
-            ampl_ts,
-            bkps_stm,
-            mean_ampl_sd,
-            sigma_ampl_sd,
-            mad_ampl_sd,
-            median_ampl_sd,
-            x_coordinates,
-            y_coordinates,
-            coordinate_type,
-        )
-
-        # Save the failed arcs to an array (such that they are not taken into account any more)
-        succeeded_arcs = [
-            arc
-            for arc, succeeded in zip(
-                arcs_updated_network, results_initial_control_network_v1["succeeded_arcs"], strict=True
-            )
-            if not np.isnan(succeeded).any()
-        ]
-
-        failed_arcs_temp = [
-            arc
-            for arc, succeeded in zip(
-                arcs_updated_network, results_initial_control_network_v1["succeeded_arcs"], strict=True
-            )
-            if np.isnan(succeeded).any() or not succeeded.any()
-        ]
-        failed_arcs = list(set(failed_arcs).union(failed_arcs_temp))
-        print(f"Failed arcs {failed_arcs}")
-
-        # Remove the failed arcs from the dictionary with all the results (as the estimated parameters etc)
-        results_initial_control_network_v2 = {}  # Make a new dictionary where we will not save
-        # the results of the failed_arcs
-
-        # Find the indices where the arcs are saved where no solution was found.
-        # These arcs have nan values
-        valid_indices = ~np.isnan(results_initial_control_network_v1["succeeded_arcs"]).any(axis=1)
-
-        # Filter all variables in the dictionary and only copy the values for the arcs where we found a solution
-        for key, value in results_initial_control_network_v1.items():
-            if (
-                isinstance(value, np.ndarray)
-                and value.shape[0] == results_initial_control_network_v1["succeeded_arcs"].shape[0]
-            ):
-                results_initial_control_network_v2[key] = value[valid_indices]
-            else:
-                results_initial_control_network_v2[key] = value  # Keep values that are not row-based unchanged
-
-        print("Check if the network with the solved arcs still meets our requirements")
-
-        # Since we have 'failed_arcs', where no solution was found, the new network need to be tested
-        # It can happen that we have isolated points,
-        # or that the average degree of the network is not high enough anymore
-        network_check, arcs_updated_network, ref_pnt = test_succeeded_arcs_control_network(
-            succeeded_arcs, quality_dict_arcs, deg_threshold, min_nodes, min_redundancy, visualize_network
-        )
-
-        if network_check == 0:
-            print("Network fails requirements, starting again")
-
-        if network_check == 1:
-            print("Network meets requirements")
-            # After the last test, the isolated arcs are removed (so they are still in
-            # 'succeeded_arcs' and in the dictionary)
-            # And these arcs need to be removed from the dictionary
-            missing_indices = [
-                idx for idx, arc in enumerate(succeeded_arcs) if arc not in arcs_updated_network
-            ]  # missing indices are the arcs that were isolated and removed. But they are still in the
-            # dictionary so there they need to be removed as well
-
-            results_control_network_temp = {}
-            for key, value in results_initial_control_network_v2.items():
-                if isinstance(value, np.ndarray):
-                    if value.ndim in {1, 2}:  # Both 1D and 2D arrays here
-                        results_control_network_temp[key] = np.delete(value, missing_indices, axis=0)
-                    else:
-                        results_control_network_temp[key] = value
-                else:
-                    results_control_network_temp[key] = value
-
-            # Now we have a network that fullfills requirements but there might be noisy arcs
-            # We compute solutions for all arcs and computed RMSE
-            # In the next part, we will remove any arcs that are too noisy for the requirements, and then test if
-            # the network still fulfills the requirements
-            print("Calculate whether there are arcs where the solution that we found is noisy")
-
-            est_displ_phase = (
-                results_control_network_temp["unwrap_phases_arc"]
-                - results_control_network_temp["estimated_cross_range_phase"]
-                - results_control_network_temp["estimated_thermal_phase"]
-            )
-            displ_phase = results_control_network_temp["estimated_displ_phase"]
-
-            residues_per_arc = est_displ_phase - displ_phase
-            sigma_post_arc = np.std(residues_per_arc, axis=1)
-            mean_sigma_prior_arc = np.mean(results_control_network_temp["sigma_phases_arc"], axis=1)
-
-            idx_bad_arcs = np.where(sigma_post_arc >= sigma_post_over_sigma_prior * mean_sigma_prior_arc)[0]
-            idx_good_arcs = np.where(sigma_post_arc < sigma_post_over_sigma_prior * mean_sigma_prior_arc)[0]
-
-            bad_arcs = results_control_network_temp["succeeded_arcs"][idx_bad_arcs]
-            good_arcs = results_control_network_temp["succeeded_arcs"][idx_good_arcs].astype(int)
-            good_arcs = [tuple(row) for row in good_arcs]  # Change the output to a list
-
-            print(f"The value for sigma_post_over_prior is {sigma_post_over_sigma_prior}")
-
-            print("we removed arcs")
-
-            print("The bad arcs are")
-            print(idx_bad_arcs)
-            print("The good arcs are")
-            print(idx_good_arcs)
-
-            # Add the noisy arcs to the list with failed_arcs
-            failed_arcs = list(set(failed_arcs).union([tuple(row.astype(int)) for row in bad_arcs]))
-            noisy_arcs = list(set(noisy_arcs).union([tuple(row.astype(int)) for row in bad_arcs]))
-
-            # Check whether the network without the noisy arcs still meets requirements
-            print("Check whether the network stil meets the requirements, even after the removal of bad arcs")
-            network_check_good, arcs_updated_network, ref_pnt = test_succeeded_arcs_control_network(
-                good_arcs, quality_dict_arcs, deg_threshold, min_nodes, min_redundancy, visualize_network=False
-            )
-
-            if network_check_good == 0:
-                print("Network does not meet requirements, start over")
-
-            if network_check_good == 1:
-                print("We are happy! The network consisting of the good arcs fulfills the requirements.")
-
-                # If there are noisy arcs, they need to be removed from the final network
-                # Convert noisy_arcs to set for quicker lookup
-                noisy_arcs_set = {tuple(map(float, arc)) for arc in noisy_arcs}
-
-                # Get the current succeeded_arcs
-                succeeded_arcs = results_control_network_temp["succeeded_arcs"]
-
-                # Deterimine which arcs to keep (so NOT in noisy_arcs_set)
-                valid_indices = np.array([tuple(row) not in noisy_arcs_set for row in succeeded_arcs])
-
-                # New dict with only valid rows
-                results_control_network = {}
-
-                for key, value in results_control_network_temp.items():
-                    if isinstance(value, np.ndarray) and value.shape[0] == succeeded_arcs.shape[0]:
-                        results_control_network[key] = value[valid_indices]
-                    else:
-                        results_control_network[key] = value  # Leave unrelated variables unchanged
-
-                network_meets_requirements = True  # Stop the loop
-
-        if not network_meets_requirements:
-            # If the network doesn't meet the requirements, start over
-            print("Network does not meet the requirements. Recomputing the network with updated failed_arcs...")
-
-    return results_control_network, ref_pnt, arcs_updated_network
-
-
-def _network_relation_matrix(idx_source, idx_target, n_points, idx_refpnt, sparse_mode: bool = False):
-    """Create the network relation matrix A as a sparse matrix.
-
-    A network relation matrix has shape (n_arcs, n_points - 1).
-    Each row corresponds to an arc, and each column corresponds to a point, excluding the reference point.
-    For each arc, the column corresponding to the source point has a value of -1, and the column corresponding
-    to the target point has a value of +1.
-    All other entries are zero.
-
-    The reference point column removal refers to Eq.4.11 of Kampes (2006).
-
-    Parameters
-    ----------
-    idx_source : list or np.ndarray
-        List of source point indices for each arc.
-    idx_target : list or np.ndarray
-        List of target point indices for each arc.
-    n_points : int
-        Total number of points in the network.
-    idx_refpnt : int
-        Index of the reference point to be excluded from the matrix. This index assumes 0-based indexing of the points.
-    sparse_mode : bool
-        Whether to return the matrix in sparse format. If False, returns a dense numpy array.
-
-    References
-    ----------
-    Kampes, Bert M. Radar interferometry: persistent scatterer technique. Dordrecht: Springer Netherlands, 2006.
-    DOI: 10.1007/978-1-4020-4723-7
-    """
-    n_arcs = len(idx_source)
-    A_start = sparse.COO(
-        (np.arange(n_arcs), idx_source),
-        np.full_like(np.arange(n_arcs), -1, dtype=np.int8),
-        shape=(n_arcs, n_points),
-    )
-    A_end = sparse.COO(
-        (np.arange(n_arcs), idx_target),
-        np.full_like(np.arange(n_arcs), 1, dtype=np.int8),
-        shape=(n_arcs, n_points),
-    )
-    A = A_start + A_end
-
-    # Convert to Compressed Sparse Row (CSR) matrix for efficient arithmetic and matrix vector operations
-    A = A.tocsr()
-
-    # Remove reference point column
-    A = scipy.sparse.hstack([A[:, :idx_refpnt], A[:, idx_refpnt + 1 :]])
-
-    if not sparse_mode:
-        A = np.array(A.todense())
-
-    return A
-
-
-def _independent_arcs(arcs: np.ndarray) -> np.ndarray:
-    """Select independent arcs from a list of arcs.
-
-    An arc is independent if its starting and ending points do not exist in any other arc's
-    starting or ending points.
-
-    Parameters
-    ----------
-    arcs : np.ndarray
-        A 2D array of shape (n_points, 2) where each row represents indices of the starting and ending points
-        of an arc.
-
-    Returns
-    -------
-    np.ndarray
-        A 2D array of independent arcs, where each row represents indices of the starting and ending points
-        of an arc.
-    """
-    # Select arcs with unique starting points
-    _, unique_idx_start = np.unique(arcs[:, 0], return_index=True)
-    arcs = arcs[unique_idx_start, :]
-
-    # Select arcs with unique ending points
-    _, unique_idx_end = np.unique(arcs[:, 1], return_index=True)
-    arcs = arcs[unique_idx_end, :]
-
-    # After previous two steps, no arcs will share starting or ending points.
-    # However, there starting points may be the ending points of other arcs, and vice versa.
-    # To ensure independency, we loop through the rest arcs and add arc one by one
-    # In each interation, remove arcs that
-    # 1) start with the ending point of this arc, or
-    # 2) end with the starting point of this arc
-    arcs_selected = np.empty((0, 2), dtype=int)
-    while arcs.shape[0] > 0:
-        arc_current = arcs[0, :]
-        arcs_selected = np.append(arcs_selected, [arc_current], axis=0)
-        # Remove arcs which contain the starting point or ending point of the current arc
-        idx_remove = np.where((arcs[:, 1] == arc_current[0]) | (arcs[:, 0] == arc_current[1]))[0]
-        # add the index of the current arc to idx_remove
-        idx_remove = np.append(idx_remove, 0)
-        arcs = np.delete(arcs, idx_remove, axis=0)
-
-    return arcs_selected
